Auto-select module clock
There are only 2 clocks for everything related to video. If one
clock is used by main display and another by video input decoder,
no clock is available for a second monitor, for example.

Instead of choosing video_pll0 clock, check if video_pll1 clock
has the desired rate. If it has, it can be shared letting video_pll0
free for other usages, so we choose it.

However, if video_pll1 can't be used, video_pll0 will be used,
posibly breaking something that was using it (like a 2nd monitor)