Fixed bugs that prevented gstreamer from working
Bug 1: in vidioc_queryctrl if control is not supported -EINVAL should
be returned, but 0 was returned instead. This caused gstreamer to
enter into an endless loop.

Bug 2: in vidioc_s_fmt_vid_cap format struct must be filled with
applied config & format, according to V4L2. Gstreamer needs some
data such as bytesperline and sizeimage.

Bug 3: bytesperline was wrong because it is suposed to be the "sprite",
that is bytes per line of the plane Y (bytes per line for plane UV are
calculated from the previous one). Instead, the driver filled bytesperline
with the average bytes per line of the whole frame, but this didn't match
the padding between lines of Y plane (which is simply the frame width).