Fix bug introduced in commit 16722b15e
Commit 16722b15e intended to fix a bug that made that all buffers
except the last one retained the first frame they were filled in the
case that buffers can't be processed quickly enough.

This fix worked but it broke non-blocking buffers dequeuing, used
for example by ffmpeg. Because of this, the driver failed to dequeue
any buffer for programs such as ffmpeg.

This commit undo commit 16722b15e and fix the bug in another way that
works for both blocking and non-blocking modes.