Dirty hack to get the address to pass it to disp driver
Demo program used for testing passed the offset returned by the
driver for each buffer to disp driver. However, after switching
from videobuf to videobuf2 this stoped working.

It was a "happy" coincidence that the offsets were the address
of the buffers too, and that's why it worked. With videobuf2 this
doesn't happen any more, and it doesn't have to do so.

As disp driver needs the physical address of the buffer, we
return this address using the reserved field of v4l2_buffer when
dequeuing a buffer. This is a very dirty hack, but disp driver
hasn't got a good design and there were not any other easy solution.

Programs like Gstreamer or ffmpeg works fine without this hack.